moved error to exceptions
