+++ flow/utils/flow_warnings.py
@@ -18,10 +18,3 @@ def deprecation_warning(obj, dep_from, dep_to):
     warnings.warn(
         "The attribute {} in {} is deprecated, use {} instead.".format(
             dep_from, obj.__class__.__name__, dep_to))
-
-
-class FatalFlowError(Exception):
-    """Exception class for Flow errors which do not allow for continuation."""
-
-    def __init__(self, msg):
-        Exception.__init__(self, msg)
