+++ tests/fast_tests/test_environment_base_class.py
@@ -9,7 +9,7 @@ from flow.controllers.car_following_models import IDMController
 from flow.controllers import RLController
 from flow.envs.loop.loop_accel import ADDITIONAL_ENV_PARAMS
 from flow.envs import Env
-from flow.utils.flow_warnings import FatalFlowError
+from flow.utils.exceptions import FatalFlowError
 
 from tests.setup_scripts import ring_road_exp_setup, highway_exp_setup
 import os
